requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
selenium==4.29.0
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
bs4
scikit-learn
apache-airflow-providers-common-compat
apache-airflow-providers-common-sql
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-odbc
apache-airflow-providers-apache-spark
fake-useragent